+++ src/ContinuityModels/ContinuityModelROF.cpp
@@ -30,17 +30,17 @@ ContinuityModelROF::ContinuityModelROF(
         u->setNoFinaicalCalculations();
     }
 
-    storage_to_rof_realization = Matrix3D<float>(
+    storage_to_rof_realization = Matrix3D<double>(
             (unsigned int) n_utilities,
             (unsigned int) NO_OF_INSURANCE_STORAGE_TIERS + 1,
             (unsigned int) ceil(WEEKS_IN_YEAR));
-    storage_to_rof_table = new Matrix3D<float>(
+    storage_to_rof_table = new Matrix3D<double>(
             (unsigned int) n_utilities,
             (unsigned int) NO_OF_INSURANCE_STORAGE_TIERS + 1,
             (unsigned int) ceil(WEEKS_IN_YEAR));
 
     /// Record which sources have no downstream sources.
-    capacities_toposorted = new float[n_topo_sources];
+    capacities_toposorted = new double[n_topo_sources];
     downstream_sources_toposort = new int[n_topo_sources];
     topo_sorted_to_all_sources = new int[n_sources];
     memset(topo_sorted_to_all_sources, -1, sizeof(int) * n_sources);
@@ -77,12 +77,12 @@ ContinuityModelROF::~ContinuityModelROF() = default;
  * Runs one the full rof calculations for realization #realization_id for a given week.
  * @param week for which rof is to be calculated.
  */
-vector<float> ContinuityModelROF::calculateShortTermROF(int week) {
+vector<double> ContinuityModelROF::calculateShortTermROF(int week) {
 
     // vector where risks of failure will be stored.
-    vector<float> risk_of_failure((unsigned long) n_utilities, 0.0);
-    vector<float> year_failure((unsigned long) n_utilities, 0.0);
-    float to_full[n_sources];
+    vector<double> risk_of_failure((unsigned long) n_utilities, 0.0);
+    vector<double> year_failure((unsigned long) n_utilities, 0.0);
+    double to_full[n_sources];
 
     for (int ws = 0; ws < n_sources; ++ws) {
         to_full[ws] = realization_water_sources[ws]->getCapacity() -
@@ -164,11 +164,11 @@ vector<float> ContinuityModelROF::calculateShortTermROF(int week) {
  * Runs one the full rof calculations for realization #realization_id for a given week.
  * @param week for which rof is to be calculated.
  */
-vector<float> ContinuityModelROF::calculateLongTermROF(int week) {
+vector<double> ContinuityModelROF::calculateLongTermROF(int week) {
 
     // vector where risks of failure will be stored.
-    vector<float> risk_of_failure((unsigned long) n_utilities, 0.0);
-    vector<float> year_failure((unsigned long) n_utilities, 0.0);
+    vector<double> risk_of_failure((unsigned long) n_utilities, 0.0);
+    vector<double> year_failure((unsigned long) n_utilities, 0.0);
 
     storage_to_rof_table->reset(NON_FAILURE);
 
@@ -212,13 +212,13 @@ vector<float> ContinuityModelROF::calculateLongTermROF(int week) {
     return risk_of_failure;
 }
 
-void ContinuityModelROF::updateStorageToROFTable(float storage_percent_decrement,
+void ContinuityModelROF::updateStorageToROFTable(double storage_percent_decrement,
                                                  int week_of_the_year,
-                                                 const float *to_full_toposort) {
-    float available_volumes[n_topo_sources];
-    float to_full[n_topo_sources];
-//    __declspec(align(64)) float spillage[n_topo_sources];
-    float spillage[n_topo_sources] __attribute__ ((aligned(64)));
+                                                 const double *to_full_toposort) {
+    double available_volumes[n_topo_sources];
+    double to_full[n_topo_sources];
+//    __declspec(align(64)) double spillage[n_topo_sources];
+    double spillage[n_topo_sources] __attribute__ ((aligned(64)));
     for (int ws = 0; ws < n_topo_sources; ++ws) {
         available_volumes[ws] = continuity_water_sources[sources_topological_order[ws]]->getAvailable_volume();
         to_full[ws] = to_full_toposort[sources_topological_order[ws]];
@@ -229,14 +229,14 @@ void ContinuityModelROF::updateStorageToROFTable(float storage_percent_decrement
     /// loops over the percent storage levels to populate table. The loop
     /// begins from one level above the level  where at least one failure was
     /// observed in the last iteration. This saves a lot of computational time.
-//    __declspec(align(64)) float delta_storage[n_topo_sources];
-//    __declspec(align(64)) float available_volumes_shifted[n_topo_sources];
-    float delta_storage[n_topo_sources] __attribute__ ((aligned(64)));
-    float available_volumes_shifted[n_topo_sources] __attribute__ ((aligned(64)));
+//    __declspec(align(64)) double delta_storage[n_topo_sources];
+//    __declspec(align(64)) double available_volumes_shifted[n_topo_sources];
+    double delta_storage[n_topo_sources] __attribute__ ((aligned(64)));
+    double available_volumes_shifted[n_topo_sources] __attribute__ ((aligned(64)));
     for (int s = beginning_tier; s <= NO_OF_INSURANCE_STORAGE_TIERS; ++s) {
         /// calculate delta storage for all reservoirs and array that will
         /// receive the shifted storage curves.
-        float percent_decrement_storage_level = (float) s * storage_percent_decrement;
+        double percent_decrement_storage_level = (double) s * storage_percent_decrement;
         std::copy(available_volumes, available_volumes + n_topo_sources, available_volumes_shifted);
 
         /// calculate the difference between the simulated available water and
@@ -253,7 +253,7 @@ void ContinuityModelROF::updateStorageToROFTable(float storage_percent_decrement
         //FIXME: ADJUST FOR TOPOLOGICAL RE-SORTING
         /// Checks for utilities failures.
         int count_fails = 0;
-        float utilities_storages[n_utilities];
+        double utilities_storages[n_utilities];
         memset(utilities_storages, 0, sizeof(int) * n_utilities);
         for (unsigned int u = 0; u < n_utilities; ++u) {
             /// Calculate combined stored volume for each utility based on
@@ -293,12 +293,12 @@ void ContinuityModelROF::updateStorageToROFTable(float storage_percent_decrement
 }
 
 //FIXME: MAKE THIS MORE EFFICIENT. THIS METHOD IS THE MOST EXPENSIVE ONE IN THE CODE.
-void ContinuityModelROF::shiftStorages(float* restrict available_volumes_shifted,
-                                       const float* restrict delta_storage,
-                                       const float* spillage) {
+void ContinuityModelROF::shiftStorages(double* restrict available_volumes_shifted,
+                                       const double* restrict delta_storage,
+                                       const double* spillage) {
 
-//    __declspec(align(64)) float available_volume_to_full[n_topo_sources];
-    float available_volume_to_full[n_topo_sources] __attribute__ ((aligned(64)));
+//    __declspec(align(64)) double available_volume_to_full[n_topo_sources];
+    double available_volume_to_full[n_topo_sources] __attribute__ ((aligned(64)));
 
     /// Add deltas to all sources following the topological order, so that
     /// upstream is calculated before downstream.
@@ -318,13 +318,13 @@ void ContinuityModelROF::shiftStorages(float* restrict available_volumes_shifted
     __assume_aligned(available_volumes_shifted, 64);
     for (int ws = 0; ws < n_topo_sources - 1; ++ws) {
         /// if not full, retrieve spill to downstream source.
-        float spillage_retrieved = min(available_volume_to_full[ws], spillage[ws])
+        double spillage_retrieved = min(available_volume_to_full[ws], spillage[ws])
                                         * (available_volume_to_full[ws] > 0);
         available_volumes_shifted[ws] += spillage_retrieved;
         available_volumes_shifted[downstream_sources_toposort[ws]] -= spillage_retrieved;
     }
     /// if not full, retrieve spill to downstream source.
-    float spillage_retrieved = min(available_volume_to_full[n_topo_sources - 1], spillage[n_topo_sources - 1])
+    double spillage_retrieved = min(available_volume_to_full[n_topo_sources - 1], spillage[n_topo_sources - 1])
                                 * (available_volume_to_full[n_topo_sources - 1] > 0);
     available_volumes_shifted[n_topo_sources - 1] += spillage_retrieved;
 }
@@ -397,7 +397,7 @@ void ContinuityModelROF::updateOnlineInfrastructure(int week) {
         }
 }
 
-const Matrix3D<float> *ContinuityModelROF::getStorage_to_rof_table() const {
+const Matrix3D<double> *ContinuityModelROF::getStorage_to_rof_table() const {
     return storage_to_rof_table;
 }
