+++ src/Controls/WwtpDischargeRule.h
@@ -16,23 +16,23 @@ private:
     vector<int> source_id_to_vector_index;
 
 public:
-    const vector<vector<float>> *year_series_fraction_discharge;
+    const vector<vector<double>> *year_series_fraction_discharge;
     const vector<int> *discharge_to_source_ids;
 
-    WwtpDischargeRule(const vector<vector<float>> *year_series_fraction_discharge,
+    WwtpDischargeRule(const vector<vector<double>> *year_series_fraction_discharge,
                       const vector<int> *discharge_to_source_ids);
 
     WwtpDischargeRule(WwtpDischargeRule &wwtp_discharge_rule);
 
     WwtpDischargeRule &operator=(const WwtpDischargeRule &wwtp_discharge_rule);
 
-    float get_dependent_variable(float water_source_id) override;
+    double get_dependent_variable(double water_source_id) override;
 
-    float get_dependent_variable(int water_source_id) override;
+    double get_dependent_variable(int water_source_id) override;
 
-    float get_dependent_variable(float x, int week) override;
+    double get_dependent_variable(double x, int week) override;
 
-    float get_dependent_variable(int x, int week) override;
+    double get_dependent_variable(int x, int week) override;
 
 };
