+++ src/DataCollector/ReservoirDataCollector.h
@@ -13,14 +13,14 @@
 class ReservoirDataCollector : public DataCollector {
 
     Reservoir *reservoir;
-    vector<float> stored_volume;
-    vector<float> total_upstream_sources_inflows;
-    vector<float> wastewater_inflows;
-    vector<float> demands;
-    vector<float> outflows;
-    vector<float> total_catchments_inflow;
-    vector<float> evaporated_volume;
-    vector<float> area;
+    vector<double> stored_volume;
+    vector<double> total_upstream_sources_inflows;
+    vector<double> wastewater_inflows;
+    vector<double> demands;
+    vector<double> outflows;
+    vector<double> total_catchments_inflow;
+    vector<double> evaporated_volume;
+    vector<double> area;
 
 public:
     explicit ReservoirDataCollector(Reservoir *reservoir);
